Profile Backend piece supposed to work now
Added so that when you enter your name, place, and pet name, it gets sent to the REST API framework. I worked on it from a separate folder that works perfectly there but when i put the changes here it doesn't work for whatever reason. I even tweaked the settings for the package.js and settings.py files and still can't get anything but the code in the views, models, serializers, and updated profile  and pagesName.js files should be working.